requirements.txt
Pillow==11.0.0
torch==2.3.0+cu121
torchvision==0.18.0+cu121
tqdm==4.66.2
